feat(polish): refactor multiadmin
